# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Slop Studios 3 API
  description: |
    AI-powered creative platform API for media generation, agent orchestration, and live coding music.

    ## Features

    - **Authentication**: JWT-based authentication with session management
    - **Claude Integration**: AI-powered text generation with streaming support
    - **Strudel Integration**: Live coding music with pattern validation and audio rendering
    - **Agent Orchestration**: Multi-agent workflows with various orchestration patterns
    - **Agent Collaboration**: Self-critique and multi-agent discussions
    - **Real-time Updates**: WebSocket support for streaming responses

    ## Authentication

    All authenticated endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    Tokens are obtained via the `/api/v1/auth/login` or `/api/v1/auth/register` endpoints.

    ## Rate Limiting

    API requests are rate-limited. When rate limited, responses include:
    - `429 Too Many Requests` status code
    - `Retry-After` header with seconds until retry is allowed

    ## Error Responses

    All error responses follow a consistent format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable error message",
        "category": "validation|authentication|authorization|not_found|conflict|rate_limit|client|external|internal|timeout",
        "details": {}
      },
...

...
servers:
...
paths:
...
  /api/v1/agents/templates:
    get:
      tags: [Templates]
      summary: List templates
      description: Returns prompt templates with optional filtering.
      operationId: listTemplates
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [system, user, workflow, helper]
        - name: tags
          in: query
          schema:
            type: string
          description: Comma-separated list of tags
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Templates retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateListResponse'

    post:
      tags: [Templates]
      summary: Create template
      description: Creates a new prompt template.
      operationId: createTemplate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or register endpoint
  schemas:
...
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Error code for programmatic handling
            message:
              type: string
              description: Human-readable error message
            category:
              type: string
              enum: [validation, authentication, authorization, not_found, conflict, rate_limit, client, external, internal, timeout]
            details:
              type: object
              additionalProperties: true
            retryAfter:
              type: integer
              description: Seconds until retry is allowed (for rate limits)
        meta:
          $ref: '#/components/schemas/ResponseMeta'

    # =========================================================================
    # Health Schemas
    # =========================================================================
...
    TemplateListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            templates:
              type: array
              items:
                $ref: '#/components/schemas/Template'
            total:
              type: integer
            page:
              type: integer
            pageSize:
              type: integer
    TemplateResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Template'
    Template:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        content:
          type: string
        variables:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
                enum: [string, number, boolean, array, object]
              required:
                type: boolean
              default: {}
              description:
                type: string
        category:
          type: string
          enum: [system, user, workflow, helper]
        tags:
          type: array
          items:
            type: string
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTemplateRequest:
      type: object
      required: [name, content, variables, category]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        content:
          type: string
          minLength: 1
          maxLength: 100000
        variables:
          type: array
          maxItems: 50
          items:
            type: object
            required: [name, type, required]
            properties:
              name:
                type: string
              type:
                type: string
                enum: [string, number, boolean, array, object]
              required:
                type: boolean
              default: {}
              description:
                type: string
        category:
          type: string
          enum: [system, user, workflow, helper]
        tags:
          type: array
          maxItems: 20
          items:
            type: string
...